# packages/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "内部 Connector 服务：catalog、配置管理、OAuth 连接与聚合 MCP。\n管理端 /admin 使用 HttpOnly SameSite=Strict Cookie；CONNECT_IT_BASE_URL 为 HTTPS 时固定设置 Secure。浏览器客户端须启用 credentials=include，且写请求须携带 X-Connect-It-CSRF: 1。Swagger 2 无法形式化表达 Cookie 鉴权。",
        "title": "connect-it API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/admin/account/password": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "修改管理员密码",
                "operationId": "changePassword",
                "parameters": [
                    {
                        "type": "string",
                        "description": "管理台同源写请求固定值 1",
                        "name": "X-Connect-It-CSRF",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "新密码（至少 8 个字符）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.changePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/connectors/{type}/auth-methods": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Connector 的认证方式与凭证字段（供创建 connection 的表单）",
                "operationId": "listAuthMethods",
                "parameters": [
                    {
                        "type": "string",
                        "description": "connector_type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.authMethodDTO"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/connectors/{type}/config": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "读取 Connector 配置（Secret 只回显已设置的 key）",
                "operationId": "getConfig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "connector_type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.configResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "写入 Connector 配置（public 全量替换、secrets 增量合并，空串删除）",
                "operationId": "putConfig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "管理台同源写请求固定值 1",
                        "name": "X-Connect-It-CSRF",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "connector_type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "配置内容；if_match 传上次读到的 updated_at（RFC3339），首次创建留空",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.putConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.configResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin"
                ],
                "summary": "删除 Connector 配置",
                "operationId": "deleteConfig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "管理台同源写请求固定值 1",
                        "name": "X-Connect-It-CSRF",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "connector_type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/connectors/{type}/config-schema": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Connector 配置表单元数据（由 Definition 的 ConfigFields 生成）",
                "operationId": "getConfigSchema",
                "parameters": [
                    {
                        "type": "string",
                        "description": "connector_type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.configFieldDTO"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/connectors/{type}/config:validate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "仅校验配置不落库",
                "operationId": "validateConfig",
                "parameters": [
                    {
                        "type": "string",
                        "description": "管理台同源写请求固定值 1",
                        "name": "X-Connect-It-CSRF",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "connector_type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "待校验配置",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.validateConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "管理员登录，成功后下发会话 cookie",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "用户名与密码",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error 是稳定、机器可读的错误码。",
                    "type": "string"
                },
                "message": {
                    "description": "Message 可安全返回调用方，不含 Provider 原始响应或 credential 值。",
                    "type": "string"
                },
                "retry_after_seconds": {
                    "description": "RetryAfterSeconds 仅在已知正数重试延迟时出现。",
                    "type": "integer"
                },
                "temporary": {
                    "description": "Temporary 表示稍后重试是否可能成功。",
                    "type": "boolean"
                }
            }
        },
...
        "api.authMethodDTO": {
            "type": "object",
            "properties": {
                "credential_fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.configFieldDTO"
                    }
                },
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "api.changePasswordRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "api.configFieldDTO": {
            "type": "object",
            "properties": {
                "default_value": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "input_type": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pattern": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "secret": {
                    "type": "boolean"
                }
            }
        },
        "api.configResponse": {
            "type": "object",
            "properties": {
                "connector_type": {
                    "type": "string"
                },
                "public": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "schema_version": {
                    "type": "integer"
                },
                "secret_keys_set": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "api.loginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.putConfigRequest": {
            "type": "object",
            "properties": {
                "if_match": {
                    "type": "string"
                },
                "public": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "secrets": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "api.validateConfigRequest": {
            "type": "object",
            "properties": {
                "public": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "secrets": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}